refactor(vault-core): expose browser-safe artifact classifier
